sockets/requirements.txt
bidict==0.23.1
click==8.1.7
h11==0.14.0
python-engineio==4.9.1
python-socketio==5.11.2
redis==5.0.6
setuptools==69.5.1
simple-websocket==1.0.0
uvicorn==0.30.1
wheel==0.43.0
wsproto==1.2.0
